Add experimental DXR optimizer
If the standard DXR optimizer fails, the experimental DXR optimizer
uses a brute-force approach to search for the required optimization
data. It is not guaranteed to produce a valid result.

+++ framework/decode/dx12_object_info.h
@@ -121,7 +121,7 @@ struct ResourceCopyInfo
 
 typedef format::ResourceValueType ResourceValueType;
 
-static uint64_t GetResourceValueSize(ResourceValueType type)
+static size_t GetResourceValueSize(ResourceValueType type)
 {
     switch (type)
     {
